+++ src/app/login/login.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from "@angular/forms";
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { CookieModule } from 'ngx-cookie';
 
 import { LoginRoutingModule } from './login-routing.module';
 import { LoginComponent } from './login.component';
@@ -13,6 +14,7 @@ import { ForgotPasswordComponent } from './component/forgot-password/forgot-pass
         LoginRoutingModule,
         ReactiveFormsModule,
         NgbModule,
+        CookieModule.forRoot(),
     ],
     declarations: [
         LoginComponent,