+++ src/app/services/domain-name.ts
@@ -1,9 +1,0 @@
-
-import { environment } from '../../environments/environment';
-
-/**
- * due to aot build const must not be declare..
- */
-export function serverDomainName() {
-	return environment.domainName;
-}