+++ src/environments/environment.ts
@@ -5,6 +5,7 @@
 
 export const environment = {
   production: false,
-  domainName: 'http://127.0.0.1:8000/api',
+  domainName: '127.0.0.1:8000/api',
+  protocol: 'http://',
   ServiceWorkersBasehref: '/bm-client/',
 };